+++ tests/test_util.py
@@ -4,7 +4,7 @@ import pytest
 import xarray as xr
 import numpy as np
 import pandas as pd
-from aqua.util import extract_literal_and_numeric, file_is_complete, to_list
+from aqua.util import extract_literal_and_numeric, file_is_complete, to_list, convert_data_units
 
 @pytest.fixture
 def test_text():
@@ -20,6 +20,30 @@ def test_extract_literal_and_numeric(test_text):
     for input_text, expected_output in test_text:
         result = extract_literal_and_numeric(input_text)
         assert result == expected_output
+
+
+@pytest.mark.aqua
+def test_convert_data_units():
+    """Test the check_data function"""
+    data = Reader(catalog='ci', model='ERA5', exp='era5-hpz3', source='monthly', loglevel=loglevel).retrieve()
+    initial_units = data['tprate'].attrs['units']
+
+    # Dataset test
+    data_test = convert_data_units(data=data, var='tprate', units='mm/day', loglevel=loglevel)
+    assert data_test['tprate'].attrs['units'] == 'mm/day'
+
+    # DataArray test
+    data = data['tprate']
+    data_test = convert_data_units(data=data, var='tprate', units='mm/day', loglevel=loglevel)
+    # We don't test values since this is done in the test_fixer.py
+    assert data_test.attrs['units'] == 'mm/day'
+    assert f"Converting units of tprate: from {initial_units} to mm/day" in data_test.attrs['history']
+
+    # Test with no conversion to be done
+    data_test = convert_data_units(data=data, var='tprate', units=initial_units, loglevel=loglevel)
+    assert data_test.attrs['units'] == initial_units
+    assert f"Converting units of tprate: from {initial_units} to mm/day" not in data_test.attrs['history']
+
 
 # Define a fixture to create a sample netCDF file for testing
 @pytest.mark.aqua
